Some change in signature

+++ include/Util.hpp
@@ -11,10 +11,11 @@
 #include <meddly.h>
 #include <vector>
 #include <iostream>
-#include "NodeFeatureVisitor.h"
 #include "ConstraintVisitor.h"
+#include "NodeFeatureVisitor.h"
 #include <fstream>
 #include "rapidxml.hpp"
+#include <gmp.h>
 
 using namespace rapidxml;
 using namespace MEDDLY;
@@ -41,10 +42,10 @@ public:
 	static dd_edge getMDDFromTuple(vector<int> tupla, forest *mdd);
 	static string* parseXML(const string &fileName);
 	static void printVector(vector<int> v, ostream &out);
-	static int getProductCountFromFile(string fileName);
-	static int getProductCountFromFile(string fileName, bool ignore);
-	static int getProductCountFromFile(string fileName, bool ignore, int reduction_factor_ctc);
-	static int getProductCountFromFile(string fileName, int reduction_factor_ctc);
+	static double getProductCountFromFile(string fileName);
+	static double getProductCountFromFile(string fileName, bool ignore);
+	static double getProductCountFromFile(string fileName, bool ignore, int reduction_factor_ctc);
+	static double getProductCountFromFile(string fileName, int reduction_factor_ctc);
 
 	static bool IGNORE_HIDDEN;
 	static bool SORT_CONSTRAINTS;
